% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vector_stats_kmers.R
\name{get_vector_stats_kmers}
\alias{get_vector_stats_kmers}
\title{Generate statistics about the kmers included in the RNAseq files}
\usage{
get_vector_stats_kmers(v_kmers_stats, fastq_path_list)
}
\arguments{
\item{v_kmers_stats}{Dataframe previously generated by the function "get_stats_kmers".}

\item{fastq_path}{path to the file containing the fastq files of the RNAseq of the vector organisms.}

\item{fastq_id_list}{list of the ID's that will be taken for the analysis. This must correspond to the name that is in the variable "fastq_path" ending for which a _1.fastq and a _2.fastq will be added after in order to acces the sequences for both directions}
}
\value{
The function will complete the "v_kmers_stats" Dataframe by adding for each kmer the frequence at which it appears in the fastq files as well as the amount of time they appeared.
}
\description{
This functions takes as imput as path to a folder or a fasta file, a k value and a x value in order to generate all k_mers that can be found in the fasta file(s) from position 0 until x.
}
\examples{
# Example of how to use the function
k_mers <- gen_obs_kmers_one_fasta("data_processed/virus/reference/Mayaro_virus.fasta", 4, 20)
print(k_mers)

}
