% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KmerEnrichFullReport.R
\name{KmerEnrichFullReport}
\alias{KmerEnrichFullReport}
\title{Generate an HTML report}
\usage{
KmerEnrichFullReport(
  virus_genome_path,
  vector_genome_path,
  k_list,
  x,
  report_name = "report.html"
)
}
\arguments{
\item{virus_genome_path}{The input is a string that can be either a path to a fasta file or a path to a folder containing fasta files.}

\item{vector_genome_path}{The input is a string that can be either a path to a fasta file or a path to a folder containing fasta files.}

\item{k_list}{A list of integers indicating the Kmers to search for in the virus genome. The report will include the Kmers that appear in all the virus genomes.}

\item{x}{Value of x indicating the x first bases in the Virus gneome in which the Kmers will be searched. The report will only keep the Kmers appearing in all the virus genomes.}

\item{report_name}{The name of the report file to generate. Default is "report.html".}
}
\value{
This function will generate an HTML report in the working directory.
}
\description{
The KmerEnrichFullReport function generates an HTML report from a template included in the package. The report includes a summary of the results of the KmerEnrich pipeline. This summary will give the main characteristics of the kmers found in the virus genome and the vector genome. The report will also include a plot of the enchrichment fo the Kmer compared in the virus versus the mosquito. The x value is the ratio of the mean count of the Kmer in the virus genome to the mean of the standard deviation of the segments separating the Kmers. The y value is the log10 of the mean frequency of the Kmer in the virus genome minus the log10 of the frequency of the Kmer in the vector genome. The report will also include a plot of the position of the Kmers in the virus genome(s).
}
