% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_kmers.R
\name{common_kmers}
\alias{common_kmers}
\title{Generate common observed kmers}
\usage{
common_kmers(fasta_folder, k, x)
}
\arguments{
\item{k}{value of the k-mer size.}

\item{x}{value indicating on how much of the sequence the kmers need to be generated.}

\item{path}{link to the fasta file (or folder containing fasta files) in string format.}
}
\value{
the function returns a list of all kmers that have been found in common in all sequences of all fasta files between position 0 and x.
}
\description{
This functions takes as imput as path to a folder or a fasta file, a k value and a x value in order to generate all k_mers that can be found in the fasta file(s) from position 0 until x.
}
\examples{
# Example of how to use the function
k_mers <- gen_obs_kmers_one_fasta("data_processed/virus/reference/Mayaro_virus.fasta", 4, 20)
print(k_mers)
k_mers <- gen_obs_kmers_one_fasta("../data_processed/virus/all_dengue/", 4, 100)
print(k_mers)

}
