% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_kmers.R
\name{common_kmers}
\alias{common_kmers}
\title{Generate common observed Kmers}
\usage{
common_kmers(virus_genome_path, k, x)
}
\arguments{
\item{virus_genome_path}{path to the folder containing the fasta files or to the fasta file. Can be a folder or a fasta file. (files must be in .fasta format).}

\item{k}{value indicating the length of the kmers that need to be searched for.}

\item{x}{value indicating the position until which the kmers need to be searched for. If x is larger than the length of the sequence, the function will take the whole sequence.}
}
\value{
The function will return a list of all kmers that have been found in common in all sequences of all fasta files between position 0 and x.
}
\description{
The common_kmers function generates a list of all kmers that have been found in common in all sequences of all fasta files between position 0 and x.
}
