% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vector_stats_kmers.R
\name{get_vector_stats_kmers}
\alias{get_vector_stats_kmers}
\title{Generate statistics about the kmers included in the RNAseq files}
\usage{
get_vector_stats_kmers(v_kmers_stats, vector_genome_path)
}
\arguments{
\item{v_kmers_stats}{Dataframe containing the kmers that need to be searched for. The dataframe must contain a column "Kmer" with the kmers that need to be searched for.}

\item{vector_genome_path}{path to the folder containing the fastq files or to the fastq file for the mosquito RNA sequences. Can be a folder or a fastq file. (files must be in .fastq format). The fastq files must end with _1.fastq and _2.fastq in order to be accessed properly. The input path on the contrary must be the part just before the _1.fastq or _2.fastq.}
}
\value{
The function will complete the "v_kmers_stats" Dataframe by adding for each kmer the frequence at which it appears in the fastq files as well as the amount of time they appeared.
}
\description{
The get_vector_stats_kmers function generates a list of all kmers that have been found in common in all sequences of all fasta files between position 0 and x.
}
