% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers_pos_df.R
\name{kmers_pos_df}
\alias{kmers_pos_df}
\title{Generate a dataframe of position of all kmers, distance between kmers for the fasta file(s)}
\usage{
kmers_pos_df(virus_genome_path, kmers_list)
}
\arguments{
\item{virus_genome_path}{The input is a string that can be either a path to a fasta file or a path to a folder containing fasta files.}

\item{kmers_list}{The input is a list of strings containing the kmers to search for.}
}
\value{
This function will return a dataframe containing the position of every kmer on every sequence of every fasta file with specific characteristics as the sequence name, the position of the kmer, the type of the virus (subtype), the segment size (distance between previous kmer and actual kmer) and the log10 value of the segment size.
}
\description{
The kmers_pos_df function is used to generate a dataframe containing the position of every kmer on every sequence of every fasta file. It also contains the size of the segments betweem 2 kmers or between the first nucleotide and the first kmer or the last kmer and the last nucleotide.
}
