% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers_pos_df.R
\name{kmers_pos_df}
\alias{kmers_pos_df}
\title{Generate a dataframe of position of all kmers, distance between kmers for the fasta file(s)}
\usage{
kmers_pos_df(path, kmers_list)
}
\arguments{
\item{path}{link to the fasta file or the folder containing fasta files in string format.}

\item{kmers_list}{List of Kmers that will be searched in the fasta file(s).}
}
\value{
the function returns a dataframe with the position of every kmer on every sequence of every fasta file. It also contains the size of the segments betweem 2 kmers or between the first nucleotide and the first kmer or the last kmer and the last nucleotide.
}
\description{
This functions takes as imput the path of a fasta file or a folder, a k value and a x value in order to generate all k_mers that can be found in the fasta file(s) from position 0 until x.
}
\examples{
# Example of how to use the function
kmer_position_df <- kmers_pos_df("../data_processed/virus/reference/Mayaro_virus.fasta", c("AAA","ACT"))
print(kmer_position_df)
kmer_position_df <- kmers_pos_df("../data_processed/virus/all_dengue/", c("AAA","ACT"))
print(kmer_position_df)

}
